Rename proto pkg to rpc
